(MIKE) reorganized monomial order code, part1, and started writing a new MonomialOrder class
