+++ M2/Macaulay2/e/Makefile.files
@@ -10,6 +10,7 @@ INTERFACE2 = interrupted
 
 INTERFACE = \
 	MonomialOrder \
+	MonomialOrderOld \
 	myalloc \
 	BasicPoly \
 	BasicPolyList \
