+++ M2/Macaulay2/e/MonomialOrder.hpp
@@ -2,128 +2,62 @@
 
 #include <string>
 #include <vector>
+#include <iostream>
 
-#include "interface/monomial-ordering.h"
+#include "interface/monomial-ordering.h" // Just for enum MonomialOrdering_type
 
-#include "ExponentVector.hpp"
+/// TODO: this is in some sense a todo list for my current modifications
+/// In Macaulay2, a MonomialOrder is a data structure that allows teh following options:
+///   - Create a monomial order from its parts
+///   - Encode a monomial into a list of ints for which comparison and multipliciation is easy.
+///   - Decode a monomial back to an exponent vector, also a "varpower" product (sparse monomial).
+///   - Comparison function
+///   - Multiplication of two monomials, with overflow checking.
+///   - to/from a "matrix" description of the monomial order.  This is lossy, in that
+///     the matrix representation doesn't know about the packing information in the order.
+/// Additionally, there are methods for querying aspects of the monomial order.
+///   - How many slots (i.e. ints in the encoded monomial) correspond to the first d parts of the monomial order?)
+///   - Is this a global ordering?
+///   - Which variables are local?
+///   - How many (and which) variables allow negative exponents?
+///   - Is this graded reverse lex order?  Is it Lex order?
+///   - What else is needed?
+/// Questions remaining:
+///   - where best to put torsion in monoids?  In the mult function?
 
-typedef struct mon_part_rec_
-{
-  enum MonomialOrdering_type type;
-  int nvars;
-  int *wts;
-} * mon_part;
-
-struct MonomialOrdering
-{
-  unsigned int _hash;
-  unsigned int len;
-  mon_part array[1];
-};
-
-class MonomialOrderings {
-public:
-  static std::string toString(const MonomialOrdering *mo);
-  static MonomialOrdering* join(const std::vector<MonomialOrdering*>& M);
-  static MonomialOrdering* product(const std::vector<MonomialOrdering*>& M);
-  static MonomialOrdering* Lex(int nvars);
-  static MonomialOrdering* Lex2(int nvars);
-  static MonomialOrdering* Lex4(int nvars);
-  static MonomialOrdering* GRevLex(int nvars);
-  static MonomialOrdering* GRevLex2(int nvars);
-  static MonomialOrdering* GRevLex4(int nvars);
-  static MonomialOrdering* GRevLex(const std::vector<int>& wts);
-  static MonomialOrdering* GRevLex2(const std::vector<int>& wts);
-  static MonomialOrdering* GRevLex4(const std::vector<int>& wts);
-  static MonomialOrdering* RevLex(int nvars);
-  static MonomialOrdering* Weights(const std::vector<int>& wts);
-  static MonomialOrdering* GroupLex(int nvars);
-  static MonomialOrdering* GroupRevLex(int nvars);
-  static MonomialOrdering* PositionUp();
-  static MonomialOrdering* PositionDown();
-
-  static MonomialOrdering* GRevLex(const std::vector<int>& wts, int packing);
-};
-
-// This is currently located in interface/monomial-ordering.cpp
-bool monomialOrderingToMatrix(
-    const struct MonomialOrdering &mo,
-    std::vector<int> &mat,
-    bool &base_is_revlex,
-    int &component_direction,      // -1 is Down, +1 is Up, 0 is not present
-    int &component_is_before_row);  // -1 means at the end, 0 means before the
-                                   // order, and r means considered before row
-                                   // 'r' of the matrix.
-
-typedef int *monomial;
-
-typedef int32_t
-    deg_t;  // this is the integer type to use for degrees and weights
-
-typedef const int *const_monomial;
-
-struct mo_block
-{
-  enum MonomialOrdering_type typ;
-  int nvars;
-  int nslots;
-  int first_exp;
-  int first_slot;
-  int nweights;
-  deg_t *weights;
-};
-
-struct MonomialOrder
-{
-  int nvars;
-  int nslots;
-  int nblocks;
-  int nblocks_before_component;
-  int nslots_before_component;
-  int component_up; /* bool */
-  deg_t
-      *degs; /* 0..nvars: heuristic degree of each variable.  degs[nvars] = 1.
-                 Assumption: degs[i] >= 1, for all i, and should be an integer.
-                 Any graded rev lex block assumes graded wrt these degrees. */
-  struct mo_block *blocks; /* 0..nblocks-1 with each entry a struct mo_block */
-  int *is_laurent; /* 0..nvars-1: 0 or 1: 1 means negative exponents allowed */
-
-public:
-  enum overflow_type { OVER, OVER1, OVER2, OVER4 } * overflow;
-  std::vector<int> overflow_flags() const;
-};
-
-MonomialOrder *monomialOrderMake(const MonomialOrdering *mo);
-void monomialOrderFree(MonomialOrder *mo);
-void monomialOrderEncodeFromActualExponents(const MonomialOrder *mo,
-                                            const_exponents a,
-                                            monomial b);
-void monomialOrderDecodeToActualExponents(const MonomialOrder *mo,
-                                          const_monomial a,
-                                          exponents_t b);
-
-std::vector<bool> laurentVariables(const MonomialOrder* mo);
-
-namespace MO { /** MO = monomial order building blocks */
+/// MO = monomial order types, internal functions and types for monomial orders */
+namespace MO {
   enum order_type {
-    Lex = 1,            /**< lexicographic order on a specified
-                             number of variables */
-    GRevLex = 4,        /**< graded reverse lexicographic order
+    Lex = 1, /**< lexicographic order on a specified number of variables */
+    GRevLex = 4, /**< graded reverse lexicographic order
                            on a specified number of variables,
-                           with (positive) weights for the grading */
-    RevLex = 10,        /**< reverse lexicographic order
+                           with weights 1 for each variable */
+    GRevLexWeights = 7, /**< graded reverse lexicographic order
+                       with specified positive weights, on a number of variables */
+    RevLex = 10, /**< reverse lexicographic order
                            on a specified number of variables.
                            Not a global order! */
     Weights = 11,
-    GroupLex = 12,    /* Lex order here */
-    GroupRevLex = 13, /* Rev lex order here */
-    Position = 16       /**< 1 or -1 for Position => Up, Position => Down.
-                           This determines the order on free modules */
+    GroupLex = 12, // previously MO_LAURENT
+    GroupRevLex = 13, // previously MO_LAURENT_REVLEX
+    PositionUp = 15,
+    PositionDown = 16
   };
+  //Previous names and values  
+  // MO_LEX = 1, 
+  // MO_GREVLEX = 4,
+  // MO_GREVLEX_WTS = 7,
+  // MO_REVLEX = 10,
+  // MO_WEIGHTS = 11,
+  // MO_LAURENT = 12,        /* Lex order here */
+  // MO_LAURENT_REVLEX = 13, /* Rev lex order here */
+  // MO_POSITION_UP = 15,
+  // MO_POSITION_DOWN = 16
 
   class Block
   {
-    order_type mType;
+  private:
+    enum order_type mType;
   
     /** number of bits per exponent/value for this type (currently 8, 16, or 32)
      *
@@ -149,16 +83,23 @@ namespace MO { /** MO = monomial order building blocks */
 
   public:
     /** constructor */
-    Block(order_type t,
+    Block(enum order_type t,
           const std::vector<int>& data,
           int& next_var,
           int& next_slot
           );
+
+    enum order_type type() const { return mType; }
+    int numVars() const { return mNumVars; }
+    auto weights() const -> const std::vector<int>& { return mWeights; }
   };
-};  // end namespace MO
+}; // end namespace MO
+
+///////////////////////////////////////////////////////////////////////
 
 class NewAgeMonomialOrder
 {
+  friend std::ostream& operator<<(std::ostream&, const NewAgeMonomialOrder& mo);
 private:
   // The sequence of blocks for the monomial order
   std::vector<MO::Block> mParts; 
@@ -180,26 +121,24 @@ private:
   std::vector<int> mIsGroupVariable;
   
 public:
-  using MOInput = std::vector<std::pair<MO::order_type, std::vector<int>>>;
+
+  using MOInput = std::vector<std::pair<enum MO::order_type, std::vector<int>>>;
   NewAgeMonomialOrder(int nvars, MOInput input);
 
   MOInput toMonomialType() const;
 
+  /// TODO
+  bool isWellDefined(int verbose_level) const;
+
+  /// TODO
+  std::ostream debugOutput(std::ostream& o) const;
 public:
   // informational functions
 public:
   // encoding/decoding/comparison of monomials.
 };
 
-bool monomialOrderingToMatrix(
-    const struct MonomialOrdering& mo,
-    std::vector<int>& mat,
-    bool& base_is_revlex,
-    int& component_direction,     // -1 is Down, +1 is Up, 0 is not present
-    int& component_is_before_row  // -1 means: at the end. 0 means before the
-                                  // order.
-    // and r means considered before row 'r' of the matrix.
-    );
+std::ostream& operator<<(std::ostream& o, const NewAgeMonomialOrder& mo);
 
 /*
 // Local Variables:
