+++ M2/Macaulay2/e/unit-tests/NCGroebnerTest.cpp
@@ -14,7 +14,7 @@
 #include "NCAlgebras/OverlapTable.hpp"
 #include "NCAlgebras/SuffixTree.hpp"
 #include "NCAlgebras/NCReduction.hpp"
-#include "MonomialOrder.hpp"
+#include "MonomialOrderOld.hpp"
 #include "monoid.hpp"
 
 #include "util-polyring-creation.hpp"
